requirements.txt
python=3.8.10
pytorch=1.9.0
torchvision=0.10.0
pandas=1.2.5